Add files via upload
IMDB dataset having 50K movie reviews for natural language processing or Text analytics. This is a dataset for binary sentiment classification containing substantially more data than previous benchmark datasets. We provide a set of 25,000 highly polar movie reviews for training and 25,000 for testing. So, predict the number of positive and negative reviews using classification.

Conclusion
As we can see after trying out all the methods, we came up with 85.7% accuracy and the accuracy out of training data is 89.7%. This shows there is a little bit of overfitting is still there in the model. We can try GridSearchCV to get the best parameter for the model. But I will continue with the accuracy of 85.7% which is quiet good looking into the training accuracy we are getting.